minor changes to handle cleaning directories more flexibly and to check cached files

+++ ketl/loader/Loader.py
@@ -49,7 +49,7 @@ class HashLoader(BaseLoader):
 
 class LocalFileLoader(BaseLoader):
 
-    def __init__(self, destination: Union[Path, str], naming_func: Callable = None, **kwargs):
+    def __init__(self, destination: Union[Path, str], naming_func: Callable = None, clean: bool = True, **kwargs):
 
         super().__init__(Path(destination))
         self.naming_func = naming_func
@@ -58,14 +58,15 @@ class LocalFileLoader(BaseLoader):
 
         # deletes compatible with pre-3.8 python
 
-        if self.destination.is_dir():
-            files = self.destination.glob('*')
-            for file in files:
-                if file.exists():
-                    file.unlink()
-        else:
-            if self.destination.exists():
+        if clean:
+            if self.destination.exists() and self.destination.is_dir():
+                files = self.destination.glob('*')
+                for file in files:
+                    if file.exists():
+                        file.unlink()
+            elif self.destination.exists():
                 self.destination.unlink()
+
 
     def full_path(self, df: pd.DataFrame):
         if not self.naming_func:
